JSON Dumping of GENERIC trees
This is the second revision of:

  https://gcc.gnu.org/pipermail/gcc-patches/2024-September/662849.html

I've incorporated the feedback given both by Richard and David - I didn't
find any memory leaks when testing in valgrind :)

As before: This patch allows the compiler to dump GENERIC trees as JSON objects.

The dump flag -fdump-tree-original-json dumps each fndecl node in the
C frontend's gimplifier as a JSON object and traverses related nodes
in an analagous manner as to raw-dumping.

Some JSON parsers expect for there to be a single JSON value per file -
the following shell command makes the output conformant:

  tr -d '\n ' < out.json | sed -e 's/\]\[/,/g' | sed -e 's/}{/},{/g'

The information in the dumped JSON is meant to be an amalgation of
tree-pretty-print.cc's dump_generic_node and print-tree.cc's debug_tree.

Bootstrapped and tested on x86_64-pc-linux-gnu without issue (again).

gcc/ChangeLog:
        * gcc/Makefile.in: Link tree-emit-json.o to c-gimplify.o
        * gcc/dumpfile.h (dump_flag): New dump flag TDF_JSON
        * gcc/tree-emit-json.cc: Logic for converting a tree to JSON
        and dumping.
        * gcc/tree-emit-json.h: For the above
gcc/c-family/ChangeLog
        * gcc/c-family/c-gimplify.cc (c_genericize): Hook for
        -fdump-tree-original-json

Signed-off-by: Thor C Preimesberger <tcpreimesberger@gmail.com>